another_function/requirements.txt
bchlib==1.0.0
easydict==1.10
kornia==0.7.0
lpips==0.1.4
matplotlib==3.2.2
numpy==1.19.5
Pillow==8.4.0
PyYAML==6.0
torchgeometry==0.1.2

absl-py                 2.1.0
bchlib                  1.0.0
cachetools              5.5.0
certifi                 2024.8.30
charset-normalizer      3.3.2
colorama                0.4.6
cycler                  0.12.1
easydict                1.10
filelock                3.16.1
fsspec                  2024.9.0
google-auth             2.35.0
google-auth-oauthlib    1.0.0
grpcio                  1.66.2
idna                    3.10
importlib_metadata      8.5.0
Jinja2                  3.1.4
kiwisolver              1.4.7
kornia                  0.7.0
lpips                   0.1.4
Markdown                3.7
MarkupSafe              2.1.5
matplotlib              3.2.2
mpmath                  1.3.0
networkx                3.1
numpy                   1.19.5
oauthlib                3.2.2
opencv-python           4.10.0.84
packaging               24.1
Pillow                  8.4.0
pip                     24.2
protobuf                5.28.2
pyasn1                  0.6.1
pyasn1_modules          0.4.1
pyparsing               3.1.4
python-dateutil         2.9.0.post0
PyYAML                  6.0
requests                2.32.3
requests-oauthlib       2.0.0
rsa                     4.9
scipy                   1.10.1
setuptools              75.1.0
six                     1.16.0
sympy                   1.13.3
tensorboard             2.14.0
tensorboard-data-server 0.7.2
torch                   1.12.0+cu113
torchgeometry           0.1.2
torchvision             0.13.0+cu113
tqdm                    4.66.5
typing_extensions       4.12.2
urllib3                 2.2.3
Werkzeug                3.0.4
wheel                   0.44.0
zipp                    3.20.2
